requirements.txt
kagglehub
sqlalchemy
psycopg2-binary
